requirements.txt

numpy
scikit-learn
torchnet
pillow
umap-learn
pandas
matplotlib
